Factor out monomorphic type checking in elaborate
Create a helper function `enforce_type` which casts an expression to a
specific type, factoring out uses of `cast` (and thereby `eq_rect`) and
also simplifying the implementation of `elaborate_unop` and
`elaborate_binop`, which now only need to use a `match` if they are
doing a polymorphic type check (e.g., checking if something is a list
but without caring what type of list).

This involves changing the types of `elaborate_unop` and
`elaborate_binop` so that they take the pre-expressions that are
arguments to the operators.

This should make it easier to prove things about `elaborate` in the
future, given that the `eq_rect` is used only as part of `enforce_type`.